xdg: consolidate migration toolkit in xcore/migrate
